+++ classes/gui/widgets/basic/progressbar.cpp
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2011 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -81,7 +81,7 @@ S::Int S::GUI::Progressbar::Paint(Int message)
 						text = String::FromInt((Int) Math::Max(value, startValue));
 						break;
 					case PB_PERCENT:
-						if (value > 0)	text = String::FromInt((Int) Math::Max(0, Math::Round(100 / ((Float) (endValue - startValue) / (Float) (value - startValue)))));
+						if (value > 0)	text = String::FromInt((Int) Math::Max(0, (Int) Math::Round(100 / ((Float) (endValue - startValue) / (Float) (value - startValue)))));
 						else		text = zeroString;
 
 						text.Append(percentString);
