+++ classes/gui/widgets/special/cursor.cpp
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2011 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -264,7 +264,7 @@ S::Int S::GUI::Cursor::Process(Int message, Int wParam, Int lParam)
 
 S::Int S::GUI::Cursor::DrawWidget()
 {
-	Int	 nMaxScrollPos = Math::Max(0, Math::Ceil((Float) (font.GetTextSizeY(text) - GetHeight()) / (font.GetTextSizeY() + 3)));
+	Int	 nMaxScrollPos = Math::Max(0, (Int) Math::Ceil((Float) (font.GetTextSizeY(text) - GetHeight()) / (font.GetTextSizeY() + 3)));
 
 	if (nMaxScrollPos != maxScrollPos)
 	{
