+++ samples/testkey/testkey.cpp
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2011 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -100,7 +100,7 @@ Int TestKey::GetBits(Int number, UnsignedInt startBit, UnsignedInt endBit)
 
 	for (UnsignedInt i = startBit; i <= endBit; i++)
 	{
-		retVal += Math::Pow(2l, i - startBit) * ((number >> i) % 2);
+		retVal += Math::Pow(2l, (Int) (i - startBit)) * ((number >> i) % 2);
 	}
 
 	return retVal;
